Changed network model and added visuals
I changed the network model and added the sigmoid activation function instead. The epochs have been increased to 25 so it will take longer to run overall. I am also training on the entire training set with a learning rate of 0.3. This allows the network to achieve over 90% accuracy in most cases. Additionally I added a performance score compared against a small sample of the test set as well as a visual representation of the actual test set image with the neural network's guess above it. I choose to run it in google colab just for simplicity and computer resources.